requirements.txt
streamlit
pandas
plotly
requests
pycoingecko
tweepy
discord.py
numpy
python-dotenv
requests
defi